Launch files overhaul (1/?) - organize launch files, view_bot.launch.py dynamic namespaces
