+++ launch/view_bot.launch.py
@@ -3,8 +3,8 @@ import os
 from ament_index_python.packages import get_package_share_directory
 
 from launch import LaunchDescription
-from launch_ros.actions import Node
-from launch.actions import IncludeLaunchDescription
+from launch.actions import IncludeLaunchDescription, DeclareLaunchArgument
+from launch.substitutions import LaunchConfiguration
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 
 import xacro
@@ -13,24 +13,37 @@ def generate_launch_description():
 
     pkg_share = get_package_share_directory('luggage_av')
 
+    namespace = LaunchConfiguration("namespace")
 
     return LaunchDescription([
+        DeclareLaunchArgument(
+            'namespace',
+            default_value="/luggage_av",
+            description="Namespace of the bot (usually its unique identifier)"
+        ), 
         IncludeLaunchDescription(
             PythonLaunchDescriptionSource([
-                os.path.join(pkg_share, "launch", "robot_state_publisher.launch.py")
-            ]),
-        ),
-        IncludeLaunchDescription(
-            PythonLaunchDescriptionSource([
-                os.path.join(pkg_share, "launch", "rviz.launch.py")
+                os.path.join(pkg_share, "launch", "nodes", "robot_state_publisher.launch.py")
             ]),
             launch_arguments=[
-                ("config_file", os.path.join(pkg_share, "rviz", "view_bot.rviz"))
+                ("namespace", namespace),
             ],
         ),
         IncludeLaunchDescription(
             PythonLaunchDescriptionSource([
-                os.path.join(pkg_share, "launch", "joint_state_publisher_gui.launch.py")
+                os.path.join(pkg_share, "launch", "nodes", "rviz.launch.py")
             ]),
+            launch_arguments=[
+                ("config_file", os.path.join(pkg_share, "rviz", "view_bot.rviz")),
+                ("namespace", namespace),
+            ],
+        ),
+        IncludeLaunchDescription(
+            PythonLaunchDescriptionSource([
+                os.path.join(pkg_share, "launch", "nodes", "joint_state_publisher_gui.launch.py")
+            ]),
+            launch_arguments=[
+                ("namespace", namespace),
+            ],
         ),
     ])
