+++ parameters/diff_drive_controller.yaml
@@ -1,4 +1,4 @@
-/luggage_av/diff_drive_controller:
+/**/diff_drive_controller:
   ros__parameters:
     type: diff_drive_controller/DiffDriveController
     left_wheel_names: ["drive_wheel_L_joint"]
